requirements.txt
asgiref==3.8.1
certifi==2024.8.30
Django==5.1.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
jalali_core==1.0.0
jdatetime==5.0.0
pillow==10.4.0
PyJWT==2.9.0
sentry-sdk==2.14.0
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
